requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
email_validator
flask_wtf